Bug 531792: [Tips] API cleanup
* Moved TipManager to internal
* Fixed all arising issues
* Added save read tips to preferences

+++ org.eclipse.tips.examples/src/org/eclipse/tips/examples/swttip/SwtTipImpl.java
@@ -44,7 +44,7 @@ public class SwtTipImpl extends Tip implements ISwtTip {
 
 	private String fSubject;
 
-	public SwtTipImpl(String providerId, int  number) {
+	public SwtTipImpl(String providerId, long number) {
 		super(providerId);
 		fSubject = "This is a tip " + number;
 	}
